msggen: Move code generation to gen.rs
