+++ src/msggen.rs
@@ -1,3 +1,4 @@
 //! Tooling to generate Rust types from ROS2 message definitions.
 pub mod parser;
 pub mod stringparser;
+pub mod generate;
